refactor(transfer): move FileCleanup transformer out of oci package
The FileCleanup transformer has no OCI-specific dependencies — it only
removes local filesystem files via file:// URIs. Move it into
transfer/internal alongside the graph helper functions that construct
the cleanup node, so all file buffer cleanup logic is co-located.

- Move FileCleanup struct and filePathFromURI from oci/transformer into
  transfer/internal/file_cleanup.go
- Move transformer tests from oci/transformer/file_cleanup_test.go into
  transfer/internal/file_cleanup_test.go
- Update transfer/internal/builder.go to reference the local FileCleanup
  instead of ocitransformer.FileCleanup
